Serialize TOML floats at the field's own precision.
Promoting float to double before std::format turned 0.1f into
0.10000000149011612, so UINode colours could not round-trip as
[0.1, 0.1, 0.12, 1.0]. AppendTOMLFloat already documents this.

+++ extras/toml/TOML.hpp
@@ -548,7 +548,7 @@ namespace detail {
         } else if constexpr (std::is_integral_v<Decayed>) {
             out += std::format("{}", value);
         } else if constexpr (std::is_floating_point_v<Decayed>) {
-            AppendTOMLFloat(out, static_cast<double>(value));
+            AppendTOMLFloat(out, value);
         } else if constexpr (StringLike<Decayed>) {
             AppendTOMLString(out, value);
         } else if constexpr (std::is_same_v<Decayed, const char*> || std::is_same_v<Decayed, char*>) {
